Working towards a first release
