--- example_files/htot_testScene_mantra1_1-24.alf
@@ -1,86 +1,0 @@
-Job -title {htot_testScene_mantra1_1-24} -priority {5} -service { PixarRender } -cleanup {
-	RemoteCmd { {TractorBuiltIn} {File} {delete} {%D(//Merlin/3d4/skid/_perso/val/htot/201911615493_htot_testScene.hipnc)} }
-	RemoteCmd { {TractorBuiltIn} {File} {delete} {%D(//Merlin/3d4/skid/_perso/val/htot/htot/htot_testScene_mantra1_1-24.alf)} }
-	RemoteCmd { {TractorBuiltIn} {File} {delete} {%D(//Merlin/3d4/skid/_perso/val/htot/ifds)} }
-	} -serialsubtasks 1 -subtasks {
-		Task {Render} -serialsubtasks 0
-		Task {IFD} -serialsubtasks 1 -cmds {
-			RemoteCmd { "hbatch.exe" -c "render -V -f 1 24 /out/mantra1; quit" //Merlin/3d4/skid/_perso/val/htot/201911615493_htot_testScene.hipnc } -service { PixarRender }
-			}
-		Task {Render Frames} -serialsubtasks 0 -subtasks {
-			Task {Render frames 1 to 24} -serialsubtasks 0 -subtasks {
-				Task {Render frame 1} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.1.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 2} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.2.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 3} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.3.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 4} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.4.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 5} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.5.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 6} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.6.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 7} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.7.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 8} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.8.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 9} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.9.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 10} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.10.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 11} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.11.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 12} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.12.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 13} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.13.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 14} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.14.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 15} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.15.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 16} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.16.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 17} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.17.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 18} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.18.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 19} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.19.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 20} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.20.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 21} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.21.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 22} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.22.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 23} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.23.ifd" } -service { PixarRender }
-				}
-				Task {Render frame 24} -cmds {
-					RemoteCmd { "mantra.exe" -V 5 -f "//Merlin/3d4/skid/_perso/val/htot/ifds/mantra1.24.ifd" } -service { PixarRender }
-				}
-			}
-		}
-	}
